Allow user overrides for tmux config
I personally have never been able to get used to this setting:

    # Renumber windows when one is deleted
    set -g renumber-windows on

However, I recognize that a number of people prefer it and that I am
probably in the minority.  This pull request (hopefully) allows for the
best of both worlds by allowing overrides in a `~/.tmux/user.conf` file.
For example, that file could contain:

    set -g renumber-windows off