Stop tracking vim spell config
A shared vim dictionary is unlikely to have a huge amount of value
moving forward, so this is me making a motion to get rid of it.  I'm
pretty sure that I'm the only one who ever committed changes to it
anyway.  :smile:

**Fair warning:** this deletes the spellings you've collected, so that
will have to be managed by individual users.  (I'm pretty sure there
would be a warning/conflict if you have changes locally.)