Make ssh_auth_sock optional
This pull request moves the `ssh_auth_sock` config into a separate
`tmux` config file that can be included in `user.conf` if the user needs
it.  This makes the default configuration appropriate for local
development, but still allows for the previous behavior if the user is
using a remote server and needing to SSH to other resources (e.g.
GitHub).

**Please note:** this means that @danbernier and @jnimety will need to add
this line to their `~/.tmux/user.conf`:

    source-file "$HOME/.tmux/ssh_auth_sock.conf"

This should resolve [@treznick's comment](https://github.com/ContinuityControl/dotfiles/issues/50#issuecomment-222540832).  Worth noting, while I can see that this file is being sourced, it's hard to test on my end.