Swift default foldmethod to index from syntax
The syntax foldmethod causes slows down vim considerably, since it has
to parse the entire file as its loading it in to detect where the folds
are. The indent foldmethod gives us much of the same benefits without
the massive slowdown.

There are perhaps other ways to solve this problem that involve keeping
the syntax foldmethod, but for now, this feels like the cleanest
approach to make our dotfiles great again as fast as possible (:troll:).